Refactor Inclusion compositional field benchmark to reduce code duplication
