+++ benchmark/inclusion/inclusion.cc
@@ -1,16 +1,4 @@
-#include <aspect/material_model/simple.h>
-#include <aspect/velocity_boundary_conditions/interface.h>
-#include <aspect/simulator_access.h>
-#include <aspect/global.h>
-
-#include <deal.II/dofs/dof_tools.h>
-#include <deal.II/numerics/data_out.h>
-#include <deal.II/base/quadrature_lib.h>
-#include <deal.II/base/function_lib.h>
-#include <deal.II/numerics/error_estimator.h>
-#include <deal.II/numerics/vector_tools.h>
-
-
+#include "inclusion.h"
 
 namespace aspect
 {
@@ -74,52 +62,19 @@ namespace aspect
       }
 
 
-
-
-
       /**
        * The exact solution for the Inclusion benchmark.
        */
       template <int dim>
-      class FunctionInclusion : public Function<dim>
+      void FunctionInclusion<dim>::vector_value (const Point< dim >   &p,
+                                                 Vector< double >   &values) const
       {
-        public:
-          FunctionInclusion (double eta_B) : Function<dim>(dim+2), eta_B_(eta_B) {}
-          virtual void vector_value (const Point< dim >   &p,
-                                     Vector< double >   &values) const
-          {
-            double pos[2]= {p(0),p(1)};
-            AnalyticSolutions::_Inclusion
-            (pos,0.2,eta_B_, &values[0], &values[1], &values[2]);
-          }
-
-        private:
-          double eta_B_;
-      };
+        double pos[2]= {p(0),p(1)};
+        AnalyticSolutions::_Inclusion
+        (pos,0.2,eta_B_, &values[0], &values[1], &values[2]);
+      }
     }
 
-
-
-    template <int dim>
-    class InclusionBoundary : public VelocityBoundaryConditions::Interface<dim>
-    {
-      public:
-        /**
-         * Constructor.
-         */
-        InclusionBoundary();
-
-        /**
-         * Return the boundary velocity as a function of position.
-         */
-        virtual
-        Tensor<1,dim>
-        boundary_velocity (const types::boundary_id ,
-                           const Point<dim> &position) const;
-
-      private:
-        double eta_B;
-    };
 
     template <int dim>
     InclusionBoundary<dim>::InclusionBoundary ()
@@ -127,16 +82,12 @@ namespace aspect
       eta_B (1e3)
     {}
 
-
-
     template <int dim>
     Tensor<1,dim>
     InclusionBoundary<dim>::
     boundary_velocity (const types::boundary_id ,
                        const Point<dim> &p) const
     {
-      Assert (dim == 2, ExcNotImplemented());
-
       double pos[2]= {p(0),p(1)};
 
       Tensor<1,dim> velocity;
@@ -146,121 +97,6 @@ namespace aspect
 
       return velocity;
     }
-
-
-
-
-    /**
-     * @note This benchmark only talks about the flow field, not about a
-     * temperature field. All quantities related to the temperature are
-     * therefore set to zero in the implementation of this class.
-     *
-     * @ingroup MaterialModels
-     */
-    template <int dim>
-    class InclusionMaterial : public MaterialModel::InterfaceCompatibility<dim>
-    {
-      public:
-        /**
-         * @name Physical parameters used in the basic equations
-         * @{
-         */
-        virtual double viscosity (const double                  temperature,
-                                  const double                  pressure,
-                                  const std::vector<double>    &compositional_fields,
-                                  const SymmetricTensor<2,dim> &strain_rate,
-                                  const Point<dim>             &position) const;
-
-        virtual double density (const double temperature,
-                                const double pressure,
-                                const std::vector<double> &compositional_fields,
-                                const Point<dim> &position) const;
-
-        virtual double compressibility (const double temperature,
-                                        const double pressure,
-                                        const std::vector<double> &compositional_fields,
-                                        const Point<dim> &position) const;
-
-        virtual double specific_heat (const double temperature,
-                                      const double pressure,
-                                      const std::vector<double> &compositional_fields,
-                                      const Point<dim> &position) const;
-
-        virtual double thermal_expansion_coefficient (const double      temperature,
-                                                      const double      pressure,
-                                                      const std::vector<double> &compositional_fields,
-                                                      const Point<dim> &position) const;
-
-        virtual double thermal_conductivity (const double temperature,
-                                             const double pressure,
-                                             const std::vector<double> &compositional_fields,
-                                             const Point<dim> &position) const;
-        /**
-         * @}
-         */
-
-        /**
-         * @name Qualitative properties one can ask a material model
-         * @{
-         */
-
-        /**
-         * Return whether the model is compressible or not.
-         * Incompressibility does not necessarily imply that the density is
-         * constant; rather, it may still depend on temperature or pressure.
-         * In the current context, compressibility means whether we should
-         * solve the contuity equation as $\nabla \cdot (\rho \mathbf u)=0$
-         * (compressible Stokes) or as $\nabla \cdot \mathbf{u}=0$
-         * (incompressible Stokes).
-         */
-        virtual bool is_compressible () const;
-        /**
-         * @}
-         */
-        /**
-         * Declare the parameters this class takes through input files.
-         */
-        static
-        void
-        declare_parameters (ParameterHandler &prm);
-
-        /**
-         * Read the parameters this class declares from the parameter file.
-         */
-        virtual
-        void
-        parse_parameters (ParameterHandler &prm);
-
-
-
-        /**
-         * @name Reference quantities
-         * @{
-         */
-        virtual double reference_viscosity () const;
-
-        virtual double reference_density () const;
-
-        virtual double reference_thermal_expansion_coefficient () const;
-
-//TODO: should we make this a virtual function as well? where is it used?
-        double reference_thermal_diffusivity () const;
-
-        double reference_cp () const;
-        /**
-         * @}
-         */
-        /**
-         * Returns the viscosity value in the inclusion
-         */
-        double get_eta_B() const;
-
-      private:
-        /**
-         * viscosity value in the inclusion
-         */
-        double eta_B;
-    };
 
     template <int dim>
     double
@@ -274,7 +110,6 @@ namespace aspect
       const double r2 = (p(0)-1.0)*(p(0)-1.0) + (p(1)-1.0)*(p(1)-1.0);
       return (r2<0.2*0.2)? eta_B : 1.0;
     }
-
 
     template <int dim>
     double
@@ -349,7 +184,6 @@ namespace aspect
       return 0;
     }
 
-
     template <int dim>
     double
     InclusionMaterial<dim>::
@@ -361,7 +195,6 @@ namespace aspect
       return 0;
     }
 
-
     template <int dim>
     double
     InclusionMaterial<dim>::
@@ -372,7 +205,6 @@ namespace aspect
     {
       return 0.0;
     }
-
 
     template <int dim>
     bool
@@ -398,8 +230,6 @@ namespace aspect
       }
       prm.leave_subsection();
     }
-
-
 
     template <int dim>
     void
@@ -431,26 +261,6 @@ namespace aspect
     }
 
 
-
-
-    /**
-      * A postprocessor that evaluates the accuracy of the solution.
-      *
-      * The implementation of error evaluators that correspond to the
-      * benchmarks defined in the paper Duretz et al. reference above.
-      */
-    template <int dim>
-    class InclusionPostprocessor : public Postprocess::Interface<dim>, public ::aspect::SimulatorAccess<dim>
-    {
-      public:
-        /**
-         * Generate graphical output from the current solution.
-         */
-        virtual
-        std::pair<std::string,std::string>
-        execute (TableHandler &statistics);
-    };
-
     template <int dim>
     std::pair<std::string,std::string>
     InclusionPostprocessor<dim>::execute (TableHandler &statistics)
@@ -478,8 +288,9 @@ namespace aspect
       Vector<float> cellwise_errors_pl2 (this->get_triangulation().n_active_cells());
 
       ComponentSelectFunction<dim> comp_u(std::pair<unsigned int, unsigned int>(0,dim),
-                                          dim+2);
-      ComponentSelectFunction<dim> comp_p(dim, dim+2);
+                                          this->get_fe().n_components());
+      ComponentSelectFunction<dim> comp_p(dim,
+                                          this->get_fe().n_components());
 
       VectorTools::integrate_difference (this->get_mapping(),this->get_dof_handler(),
                                          this->get_solution(),
